+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 - **Enhancement**: Check Kotlin version compatibility. Use the `metro.version.check=false` Gradle property to disable these warnings if you're feeling adventurous.
 - **Fix**: Fix class-private qualifiers on multibinding contributions in other modules not being recognized in downstream graphs.
 - **Fix**: Fix member injectors not getting properly visited in graph validation.
+- **Change**: Change `InstanceFactory` to a value class.
 
 0.1.1
 -----